# openapi/harness-core-v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "auto-agent HTTP API",
    "version": "v1",
    "description": "Versioned SaaS HTTP/SSE contract for the routes registered by pkg/server. Tenant and subject ownership are resolved from the authenticated Principal; cross-tenant session access is intentionally reported as 404. This pre-GA v1 contract accepts additive fields and routes, while semantic or breaking changes require a new reviewed contract revision. Operations marked x-harness-contract-first are approved control-plane contracts whose server registration is verified before they can be considered implemented. Optional adapters return the documented 501 error when they are not configured."
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/v1/sessions/{id}/runs": {
      "get": {"operationId": "listRuns", "tags": ["runs"], "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/SessionID"}], "responses": {"200": {"description": "durable runs for the session", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunList"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "501": {"$ref": "#/components/responses/Error"}}},
      "post": {"operationId": "runTurn", "tags": ["runs"], "summary": "Run one synchronous turn as server-sent events", "security": [{"BearerAuth": []}], "x-harness-reconnect": "The SSE stream has no event IDs, Last-Event-ID support, or server-side replay. On disconnect, use GET /v1/sessions/{id}/events with after_seq to retrieve durable events, then inspect GET run state when configured.", "parameters": [{"$ref": "#/components/parameters/SessionID"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunRequest"}}}}, "responses": {"200": {"description": "SSE stream. Event names are persisted session event types; control/store errors use the same Error payload.", "content": {"text/event-stream": {"schema": {"$ref": "#/components/schemas/SSEEvent"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}
    },
...
    "/v1/admin/obs/rules": {"get": {"operationId": "listObservationRules", "tags": ["admin"], "security": [{"BearerAuth": []}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}, "post": {"operationId": "createObservationRule", "tags": ["admin"], "security": [{"BearerAuth": []}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}},
    "/v1/admin/obs/rules/{id}": {"put": {"operationId": "updateObservationRule", "tags": ["admin"], "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/ID"}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}, "delete": {"operationId": "deleteObservationRule", "tags": ["admin"], "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/ID"}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}},
    "/v1/admin/obs/hits": {"get": {"operationId": "listObservationHits", "tags": ["admin"], "security": [{"BearerAuth": []}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}},
    "/v1/admin/backtests": {"post": {"operationId": "runBacktest", "tags": ["admin"], "security": [{"BearerAuth": []}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}},
...
    "/v1/admin/tool-library/observations": {"get": {"operationId": "listToolLibraryObservations", "tags": ["admin"], "security": [{"BearerAuth": []}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {"type": "http", "scheme": "bearer", "description": "Built-in account token, or a deployment-injected Authenticator. The resolved Principal supplies tenant, subject, scope, grants, and role; handlers enforce ownership and administrator boundaries. Development header fallback is disabled in production."},
      "RunnerBearer": {"type": "http", "scheme": "bearer", "description": "Private runner token; it cannot access administrative runner task routes."}
    },
    "parameters": {
      "SessionID": {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}},
...
      "ID": {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}},
...
    },
    "responses": {
      "Success": {"description": "Implemented successful response; exact payload varies by operation."},
      "Error": {"description": "Uniform JSON error envelope", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
      "Error": {"type": "object", "required": ["error"], "properties": {"error": {"type": "string"}}, "additionalProperties": false},
...
      "RunRequest": {"type": "object", "required": ["message"], "properties": {"message": {"type": "string", "minLength": 1}}, "additionalProperties": true},
      "RunRecord": {"type": "object", "required": ["run_id", "session_id", "status"], "properties": {"run_id": {"type": "string"}, "session_id": {"type": "string"}, "status": {"type": "string"}, "attempt": {"type": "integer"}, "generation": {"type": "integer"}}},
      "RunList": {"type": "object", "required": ["runs"], "properties": {"runs": {"type": "array", "items": {"$ref": "#/components/schemas/RunRecord"}}}},
...
      "SSEEvent": {"type": "string", "description": "SSE wire frame: event: <session event type> followed by data: <JSON>. Examples include run/start, assistant/chunk, tool/call, approval/requested, run/end, store/error, and control/error."},
...
    }
  }
}